+++ src/com/intellij/idea/plugin/hybris/impex/actions/ImpExTableSelectAction.kt
@@ -27,6 +27,7 @@ import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiFile
 import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.util.concurrency.AppExecutorUtil
 
@@ -40,7 +41,7 @@ class ImpExTableSelectAction : AbstractImpExTableAction() {
         }
     }
 
-    override fun performAction(project: Project, editor: Editor, element: PsiElement) {
+    override fun performAction(project: Project, editor: Editor, psiFile: PsiFile, element: PsiElement) {
         ReadAction
             .nonBlocking<TextRange?> {
                 return@nonBlocking when (element) {