+++ src/com/intellij/idea/plugin/hybris/impex/actions/ImpExTableSplitVerticallyAction.kt
@@ -24,6 +24,7 @@ import com.intellij.openapi.command.WriteCommandAction
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiFile
 import com.intellij.psi.codeStyle.CodeStyleManager
 import com.intellij.psi.impl.source.PostprocessReformattingAspect
 import com.intellij.psi.util.childrenOfType
@@ -49,7 +50,7 @@ class ImpExTableSplitVerticallyAction : AbstractImpExTableColumnAction() {
         }
     }
 
-    override fun performAction(project: Project, editor: Editor, element: PsiElement) {
+    override fun performAction(project: Project, editor: Editor, psiFile: PsiFile, element: PsiElement) {
         val headerParameter = when (element) {
             is ImpexFullHeaderParameter -> element
             is ImpexValueGroup -> element.fullHeaderParameter